6.0_Assignment/requirements.txt

streamlit
pandas
sentence-transformers
faiss-cpu
openai
numpy
openai>=1.0.0